# openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "negative25 API", "version": "1.0.0", "description": "Versioned photography archive API shared by Web and future iPhone clients." },
  "servers": [{ "url": "/api/v1" }],
  "paths": {
...
    "/auth/login": { "post": { "summary": "Create a session", "responses": { "200": { "description": "Access and refresh tokens" }, "401": { "description": "Invalid credentials" } } } },
...
    "/workspaces": { "get": { "summary": "List workspaces available to the current user", "responses": { "200": { "description": "Workspace memberships" }, "401": { "description": "Authentication required" } } } },
    "/spaces/{slug}": { "get": { "summary": "Read a public space", "parameters": [{ "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Space" }, "404": { "description": "Not found" } } } },
    "/spaces/{slug}/photos": { "get": { "summary": "List published photos", "parameters": [{ "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }, { "name": "mode", "in": "query", "schema": { "type": "string", "enum": ["featured", "recent", "shuffle", "location", "nearby", "faraway"] } }, { "name": "location", "in": "query", "description": "Stable region ID or normalized location slug", "schema": { "type": "string" } }, { "name": "cursor", "in": "query", "schema": { "type": "string" } }, { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 24 } }], "responses": { "200": { "description": "Cursor-paginated photo list" } } } },
...
    "/photos/{id}": { "get": { "summary": "Read one published photo", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Photo" }, "404": { "description": "Not found" } } } },
...
    "/admin/spaces/{slug}/photos/{id}": { "patch": { "summary": "Update photo visibility and metadata", "parameters": [{ "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }, { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Updated photo" } } } },
...
  }
}